feat(dropdown-menu): add no-result menu item
The changes add a new `NoContentItem` composable function to the `DropdownMenu` component. This function is used to display a menu item with a centered, italicized text when there are no results to display in the dropdown menu.

The changes also include the addition of a new string resource `spark_dropdown_menu_item_no_result_label` to provide the default text for the no-result menu item.

This feature is important to provide a clear indication to the user when the dropdown menu does not contain any items, improving the overall user experience.